requirements.txt
requests>=2.26.0
pandas>=1.3.3
psycopg2-binary>=2.9.1
python-dotenv>=0.19.1
flask>=2.0.0
flask-cors>=4.0.0